Support single-table formats in `read_tables`/`write_tables`
This allows you to call the `write_tables` function with a file
specification that only supports a single table, as long as you've only
got a single table to write. And similarly for `read_tables`.

This allows us to use a common code path for datasets which will only
produce a single results table and those which will produce multiple
(and similarly will require multiple dummy data files).

We pay special attention to making the error messages here as helpful as
possible. This is on the assumption that the first time anyone tries to
use event-level data they won't immediately realise that they need to
specify their output and dummy data files in a different way.