feat: refactor measures calculation to use in-sql summing and grouping
Replace the fetchers with a simple tuple of sum-over and group-by column
indexes for each measure, and pass this to the query engine to fetch
results.

We make use of the new ability to return multiple results tables, and
return one table per measure. The rows in the results are now already
summed and grouped.