Add a test for measures where the interval population is dependent on the specific interval
Tests the case where an interval denominator is dependent on the specific
interval, and the union of all denominators would not capture the full
population.

If we want to extract variables that don't vary with the interval (e.g.
sex, ethnicity), we can get the variable definition for them, but we'd
need a population to extract it for. We could define a population that
covers the entire interval period, from first start date to first end
date, which would work in many cases, but would not work if the
interval population is dependent on the specific interval in some way.
This test would catch that erroneous assumption.